+++ app/jobs/submission_ingestion_job.rb
@@ -1,0 +1,6 @@
+class SubmissionIngestionJob < ApplicationJob
+  def perform(submission_file)
+    orchestrator = Ingest::Orchestrator.new(submission_file)
+    orchestrator.perform
+  end
+end