+++ app/models/framework/definition/ast/presenter.rb
@@ -38,8 +38,7 @@ class Framework
         end
 
         def field_by_name(entry_type, name)
-          field_def = field_defs(entry_type).find { |f| f[:field] == name }
-          raise ArgumentError, "No #{entry_type} field '#{name}' found" unless field_def
+          field_def = field_defs(entry_type).find { |f| f[:field] == name } || return
 
           Field.new(field_def, lookups)
         end