+++ spec/features/admin_can_edit_a_framework_spec.rb
@@ -1,6 +1,11 @@
 require 'rails_helper'
 
 RSpec.feature 'Admin can edit a framework' do
+  let(:framework) do
+    create(:framework, published: published, short_name: 'RM999',
+      name: 'Framework to be changed', definition_source: existing_source)
+  end
+
   let(:existing_source) do
     <<~FDL
       Framework RM999 {
@@ -17,7 +22,7 @@ RSpec.feature 'Admin can edit a framework' do
     # Given that I am logged in as an admin
     sign_in_as_admin
     # And there is an existing framework
-    create(:framework, short_name: 'RM999', name: 'Framework to be changed', definition_source: existing_source)
+    framework
   end
 
   context 'we have valid framework source' do
@@ -33,7 +38,9 @@ RSpec.feature 'Admin can edit a framework' do
       FDL
     end
 
-    context 'there is an existing framework' do
+    context 'there is an existing unpublished framework' do
+      let(:published) { false }
+
       scenario 'everything is fine' do
         # When I visit the frameworks page
         visit admin_frameworks_path
@@ -57,6 +64,35 @@ RSpec.feature 'Admin can edit a framework' do
         expect(page).to have_content('Vehicle Leasing')
       end
     end
+
+    context 'there is an existing published framework' do
+      let(:published) { true }
+
+      scenario 'no link exists to edit' do
+        # When I visit the frameworks page
+        visit admin_frameworks_path
+        # Then I should see a list of frameworks
+        expect(page).to have_content('Framework to be changed')
+        # When I click on the existing framework
+        click_link('Framework to be changed')
+        # Then I should see the framework
+        expect(page).to have_content('Framework RM999 {')
+        # And I should not see 'Edit definition'
+        expect(page).not_to have_content('Edit definition')
+      end
+
+      scenario 'an admin attempts to /edit manually' do
+        # When I try to edit the framework
+        visit edit_admin_framework_path(framework)
+
+        # Then I should be redirected to the list of frameworks
+        expect(page).to have_content('Frameworks')
+        expect(page).to have_content('New Framework')
+
+        # And I should see that I cannot edit the framework
+        expect(page).to have_content('This framework has been published and cannot be edited')
+      end
+    end
   end
 
   context 'we have invalid framework source' do
@@ -72,7 +108,9 @@ RSpec.feature 'Admin can edit a framework' do
       FDL
     end
 
-    context 'there is an existing framework' do
+    context 'there is an existing unpublished framework' do
+      let(:published) { false }
+
       scenario 'the existing framework is not updated' do
         # When I visit the frameworks page
         visit admin_frameworks_path